+++ src/platforms.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 Comcast Cable Communications Management, LLC
+ * Copyright 2023 Comcast Cable Communications Management, LLC
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -15,10 +15,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import Circle from './Circle.js'
-import RouterView from './RouterView.js'
-
-export default () => ({
-  Circle: Circle(),
-  RouterView: RouterView(),
-})
+export {
+  WebPlatform,
+  WebPlatformNext,
+  WebPlatformChrome50,
+  WebPlatformLegacy,
+} from '@lightningjs/renderer/platforms'